Modulo IV - Ciencia Da Computação/bloco-33/aula-33.3/exercicios-bloco/tests/test_exercicios_bloco_4.py
from src.exercicios_bloco_4 import filter_valid_emails


def test_if_dont_have_emails_returns_empty_list():
    assert filter_valid_emails([]) == []


def test_only_valid_emails():
    emails = ["username@website.com"]
    expected_emails = ["username@website.com"]
    assert filter_valid_emails(emails) == expected_emails


def test_invalid_emails_should_be_filtered():
    emails = ["inv*alid@website.com"]
    expected_emails = []
    assert filter_valid_emails(emails) == expected_emails


def test_valid_and_invalid_emails_returns_only_valids():
    emails = ["username@website.com", "inv*alid@website.com"]
    expected_emails = ["username@website.com"]
    assert filter_valid_emails(emails) == expected_emails
